requirements.txt
streamlit
pandas
google-cloud-bigquery
db-dtypes
plotly
numpy
scipy